Move all tests to pytest style, extract datastore storage and broadcast queue
